FritzAsephRuaro_input&output.cpp

+++ main.cpp
@@ -4,9 +4,10 @@ using namespace std;
 
 int main(int argc, char** argv) {
 	
-	float num1, num2, num3;
+	double num1, num2, num3;
 	cout<<"Enter three numbers:";
 	cin >>num1>>num2>>num3;
+	
 	double sum=num1 + num2 + num3;
 	cout<<"Addition:"<<sum<<endl;
